requirements.txt
langchain
langchain-google-genai
langgraph
psycopg2-binary
python-dotenv
pandas
tabulate